refactor(mbforge-domain): import order + minor cleanup
Pure mechanical pass on the 8 files NOT touched by the zvec migration:
- Reorder `use` declarations (crate-local first)
- Collapse/expand expressions to rustfmt defaults
- Remove unnecessary cast in content_cache.rs::trim
- Collapse long match arms in project.rs::open
- resource_manager.rs had the most touch points (multi-line closures,
  log reformat)

+++ src-tauri/crates/mbforge-domain/src/document/content_cache.rs
@@ -121,7 +121,7 @@ impl ContentCache {
                     ORDER BY hit_count ASC, created_at ASC
                     LIMIT ?1
                 )",
-                params![(count - self.max_entries as i64 + 1) as i64],
+                params![count - self.max_entries as i64 + 1],
             )
             .ok();
         }
